requirements.txt
asgiref==3.6.0
Django==4.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
